timeouts now work

+++ starboard.py
@@ -9,12 +9,16 @@
 # this means we have to make sure every user can't star any message more than once (hence the UNIQUE(starrer,msg) below)
 # also, to not get strange behaviour like fake stars or double counts, when a star is removed, it has to be removed in
 #   the same medium as it was added.
+#
+# timeout_d (interval in days) is optional. after timeout_d passes from the message being sent, the message will be
+#   locked in its awarded/unawarded state. however, star counts are still updated.
 
 SCHEMA = f"""BEGIN;
 CREATE TABLE IF NOT EXISTS guilds(
     guild   INTEGER PRIMARY KEY,
     minimum INTEGER NOT NULL DEFAULT 3,
-    sb      INTEGER NOT NULL UNIQUE  -- starboard channel
+    sb      INTEGER NOT NULL UNIQUE, -- starboard channel
+    timeout INTEGER                  -- timeout in days (NULL means no timeout)
 );
 CREATE TABLE IF NOT EXISTS awarded(
     msg     INTEGER PRIMARY KEY,     -- original message
@@ -39,6 +43,7 @@ import discord.ext.commands as commands
 import aiosqlite
 import asyncio
 import logging
+import datetime
 
 def calc_color(count:int) -> discord.Colour:
     return discord.Colour.from_rgb(255,255,max(0,min(255,1024//(count+4)-20)))
@@ -48,9 +53,12 @@ def build_message(count:int, msg:discord.Message) -> dict:
             "embed": discord.Embed(colour=calc_color(count), description=msg.content)
                      .set_author(name=msg.author.display_name, icon_url=msg.author.display_avatar.url) }
 
-def ephemeral(c,*args,**kwargs): return c.response.send_message(*args,ephemeral=True,**kwargs)
+def ephemeral(c, *args, **kwargs): return c.response.send_message(*args, ephemeral=True, **kwargs)
 
-commands.MessageConverter
+def on_time(msg_id:int, timeout_d:int) -> bool:  # True if the timeout hasn't passed yet
+    if timeout_d is None: return True
+    send_time = discord.utils.snowflake_time(msg_id)
+    return datetime.datetime.now(datetime.UTC) < send_time + datetime.timedelta(minutes=timeout_d)
 
 class NotConfigured(Exception): pass
 
@@ -91,22 +99,37 @@ class Starboard(commands.Cog):
         await ctx.send(**build_message(count, await self.partial_msg(msg_ch_id,msg_id).fetch()))
 
     @app_commands.command(description="change configuration like msg_sb channel or min stars")
-    @app_commands.rename(sb="starboard-channel",minimum="minimum-star-count")
+    @app_commands.rename(sb="starboard-channel", minimum="minimum-star-count", timeout_d="timeout-in-days")
     @app_commands.default_permissions(manage_channels=True)
-    async def starconfig(self, c:discord.Interaction, sb:discord.TextChannel|None=None, minimum:int|None=None):
-        if sb is None and minimum is None:
-            await self.db.execute("DELETE FROM guilds WHERE guild=?", (c.guild_id,))
-            await ephemeral(c,"unconfigured")
+    async def starconfig(self, c:discord.Interaction,
+                         sb:discord.TextChannel|None=None, minimum:int|None=None, timeout_d:int|None=None):
+        if sb is None and minimum is None and timeout_d is None:
+            match await self.db_fetchone("SELECT minimum,sb,timeout FROM guilds WHERE guild=?", (c.guild_id,)):
+                case minimum,sb_id,None:
+                    msg = f"minimum stars: {minimum}\nstarboard channel: <#{sb_id}>\ntimeout: never"
+                case minimum,sb_id,timeout_d:
+                    msg = f"minimum stars: {minimum}\nstarboard channel: <#{sb_id}>\ntimeout: {timeout_d} days"
+                case None:
+                    msg = "unconfigured"
+            return await ephemeral(c, msg)
         else:
             if sb is not None:  # do this first in case minimum isn't set, bc that has a minimum value
                 if sb.guild != c.guild: return await ephemeral(c,"eat bricks")
-                await self.db.execute("INSERT OR REPLACE INTO guilds(guild,sb) VALUES(?,?)", (c.guild_id, sb.id))
+                cur = await self.db.execute("INSERT OR REPLACE INTO guilds(guild,sb) VALUES(?,?)", (c.guild_id, sb.id))
             if minimum is not None:
                 if minimum < 1: return await ephemeral(c,"eat bricks")
                 cur = await self.db.execute("UPDATE guilds SET minimum=? WHERE guild=?", (minimum, c.guild_id))
                 if cur.rowcount==0:
                     await self.db.rollback()
                     return await ephemeral(c, "no starboard channel set")
+            if timeout_d is not None:
+                if timeout_d == 0: timeout_d = None
+                elif timeout_d < 0: return await ephemeral(c,"eat bricks")
+                cur = await self.db.execute("UPDATE guilds SET timeout=? WHERE guild=?", (timeout_d, c.guild_id))
+                if cur.rowcount==0:
+                    await self.db.rollback()
+                    return await ephemeral(c, "no starboard channel set")
+
         await ephemeral(c, "ok")
         await self.db.commit()
 
@@ -139,10 +162,11 @@ class Starboard(commands.Cog):
 
     async def get_guild_info(self, guild_id:int) -> dict:
         try:
-            minimum,sb_id = await self.db_fetchone("SELECT minimum,sb FROM guilds WHERE guild=?", (guild_id,))
+            minimum,sb_id,timeout_d = await self.db_fetchone(
+                "SELECT minimum,sb,timeout FROM guilds WHERE guild=?", (guild_id,))
         except TypeError as e:
             raise NotConfigured() from e
-        return {"minimum":minimum, "sb_id":sb_id, "guild_id":guild_id}
+        return {"minimum":minimum, "sb_id":sb_id, "timeout_d":timeout_d, "guild_id":guild_id}
 
     async def find_msg(self, minimum:int, sb_id:int, msg_id:int, msg_ch_id:int, guild_id:int,
                        msg:discord.Message|None=None, **_) -> dict:
@@ -155,8 +179,9 @@ class Starboard(commands.Cog):
             except TypeError: pass  # message in starboard but not managed by this bot. i'll allow it
         return {"msg_id":msg_id, "msg_ch_id":msg_ch_id, "medium":medium, "msg":msg}
     
-    async def add_star(self, minimum:int, sb_id:int, msg_id:int, msg_ch_id:int, guild_id:int, user_id:int, medium:int,
-                       msg:discord.Message|None=None) -> bool:  # return True if the star was there already
+    async def add_star(self, minimum:int, sb_id:int, timeout_d:int|None, msg_id:int, msg_ch_id:int, guild_id:int,
+                       user_id:int, medium:int, msg:discord.Message|None=None
+                      ) -> bool:  # return True if the star was there already
         if (await self.db.execute("INSERT OR IGNORE INTO stars(starrer,msg,guild,medium) VALUES(?,?,?,?)",
                                   (user_id,msg_id,guild_id,medium))).rowcount == 0:
             return False
@@ -168,20 +193,22 @@ class Starboard(commands.Cog):
         match await self.db_fetchone("SELECT msg_sb FROM awarded WHERE msg=?", (msg_id,)):
             case msg_sb_id,:  # already in starboard, edit the message
                 await self.partial_msg(sb_id,msg_sb_id).edit(**build_message(count, msg))
-            case None:        # not in starboard yet (usually bc count==minimum, or minimum was higher back then)
+            case None if on_time(msg_id,timeout_d):
+                # not in starboard yet (usually bc count==minimum, or minimum was higher back then)
                 msg_sb = await self.bot.get_channel(sb_id).send(**build_message(count, msg))
                 await self.db.execute("INSERT INTO awarded(msg,msg_sb,msg_ch,guild,author) VALUES(?,?,?,?,?)",
                                       (msg_id, msg_sb.id, msg_ch_id, guild_id, msg.author.id))
         await self.db.commit()
         return True
     
-    async def remove_star(self, msg_id:int, msg_ch_id:int, guild_id:int, user_id:int, minimum:int, sb_id:int, medium:int,
-                          msg:discord.Message|None=None) -> bool:
+    async def remove_star(self, minimum:int, sb_id:int, timeout_d:int|None, msg_id:int, msg_ch_id:int, guild_id:int,
+                          user_id:int, medium:int, msg:discord.Message|None=None
+                         ) -> bool:
         dlt = await self.db.execute("DELETE FROM stars WHERE starrer=? AND msg=? AND medium=?", (user_id, msg_id, medium))
         if dlt.rowcount == 0:
             return False # don't bother continuing if the star wasn't recorded or in a different medium
         count, = await self.db_fetchone("SELECT count(starrer) FROM stars WHERE msg=?", (msg_id,))
-        if count<minimum:  # message unawarded, or it wasn't awarded to begin with
+        if count<minimum and on_time(msg_id,timeout_d):  # message unawarded, or it wasn't awarded to begin with
             match await self.db_fetchone("DELETE FROM awarded WHERE msg=? RETURNING msg_sb", (msg_id,)):
                 case msg_sb_id,: await self.partial_msg(sb_id,msg_sb_id).delete()
         else:  # unstarred, but the message can stay in starboard
@@ -189,8 +216,9 @@ class Starboard(commands.Cog):
             match await self.db_fetchone("SELECT msg_sb FROM awarded WHERE msg=?", (msg_id,)):
                 case msg_sb_id,:
                     await self.partial_msg(sb_id,msg_sb_id).edit(**build_message(count, msg))
-                case None:  # edge case: the message was and still is award-worthy, but it wasn't sent (probably because
-                            # minimum was higher). we add it anyways, to be consistent with star add
+                case None if on_time(msg_id,timeout_d):
+                    # edge case: the message was and still is award-worthy, but it wasn't sent (probably because minimum
+                    # was higher). we add it anyways, to be consistent with star add
                     msg_sb = await self.bot.get_channel(sb_id).send(**build_message(count, msg))
                     await self.db.execute("INSERT INTO awarded(msg,msg_sb,msg_ch,guild,author) VALUES(?,?,?,?,?)",
                                           (msg_id, msg_sb.id, msg_ch_id, guild_id, msg.author.id))
